examples/led/requirements.txt
gpiozero==1.5.1
paho-mqtt==1.5.1
